requirements.txt
openai
google-genai
chromadb
pymupdf
python-docx
python-dotenv
fastapi
uvicorn
